splitting codes before sdxl support
