+++ merge_PermSpec_SDXL.py
@@ -1,0 +1,1260 @@
+# https://github.com/vladmandic/automatic/blob/dev/modules/merging/merge_PermSpec_SDXL.py
+from weight_matching import PermutationSpec, permutation_spec_from_axes_to_perm
+
+def sdxl_permutation_spec() -> PermutationSpec:
+    conv = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        f"{name}.weight": (
+            p_out,
+            p_in,
+        ),
+        f"{name}.bias": (p_out,),
+    }
+    norm = lambda name, p: {f"{name}.weight": (p,), f"{name}.bias": (p,)}  # pylint: disable=unnecessary-lambda-assignment
+    dense = (
+        lambda name, p_in, p_out, bias=True: {  # pylint: disable=unnecessary-lambda-assignment
+            f"{name}.weight": (p_out, p_in),
+            f"{name}.bias": (p_out,),
+        }
+        if bias
+        else {f"{name}.weight": (p_out, p_in)}
+    )
+    # Layer name is somehow changed!
+    dense2 = (
+        lambda name, p_in, p_out, bias=True: {  # pylint: disable=unnecessary-lambda-assignment
+            f"{name}_weight": (p_out, p_in),
+            f"{name}_bias": (p_out,),
+        }
+        if bias
+        else {f"{name}_weight": (p_out, p_in)}
+    )
+    skip = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        f"{name}": (
+            p_out,
+            p_in,
+            None,
+            None,
+        )
+    }
+
+    # Unet Res blocks
+    easyblock = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        **norm(f"{name}.in_layers.0", p_in),
+        **conv(f"{name}.in_layers.2", p_in, f"P_{name}_inner"),
+        **dense(
+            f"{name}.emb_layers.1", f"P_{name}_inner2", f"P_{name}_inner3", bias=True
+        ),
+        **norm(f"{name}.out_layers.0", f"P_{name}_inner4"),
+        **conv(f"{name}.out_layers.3", f"P_{name}_inner4", p_out),
+    }
+
+    return permutation_spec_from_axes_to_perm(
+        {
+            # Skipped Layers
+            **skip("betas", None, None),
+            **skip("alphas_cumprod", None, None),
+            **skip("alphas_cumprod_prev", None, None),
+            **skip("sqrt_alphas_cumprod", None, None),
+            **skip("sqrt_one_minus_alphas_cumprod", None, None),
+            **skip("log_one_minus_alphas_cumprods", None, None),
+            **skip("sqrt_recip_alphas_cumprod", None, None),
+            **skip("sqrt_recipm1_alphas_cumprod", None, None),
+            **skip("posterior_variance", None, None),
+            **skip("posterior_log_variance_clipped", None, None),
+            **skip("posterior_mean_coef1", None, None),
+            **skip("posterior_mean_coef2", None, None),
+            **skip("log_one_minus_alphas_cumprod", None, None),
+            **skip("model_ema.decay", None, None),
+            **skip("model_ema.num_updates", None, None),
+            **skip("conditioner.embedders.0.transformer.text_model.embeddings.position_ids", None, None),
+            **skip("conditioner.embedders.1.model.logit_scale", None, None),
+            **skip("conditioner.embedders.1.model.positional_embedding", None, None),
+            **skip("conditioner.embedders.1.model.text_projection", None, None),
+            **conv("model.diffusion_model.input_blocks.0.0", "P_bg0", "P_bg1"),
+            **easyblock("model.diffusion_model.input_blocks.1.0", "P_bg2", "P_bg3"),
+            **easyblock("model.diffusion_model.input_blocks.2.0", "P_bg4", "P_bg5"),
+            **conv("model.diffusion_model.input_blocks.3.0.op", "P_bg6", "P_bg7"),
+            **easyblock("model.diffusion_model.input_blocks.4.0", "P_bg8", "P_bg9"),
+            **conv("model.diffusion_model.input_blocks.4.0.skip_connection", "P_bg10", "P_bg11"),
+            **norm("model.diffusion_model.input_blocks.4.1.norm", "P_bg12"),
+            **conv("model.diffusion_model.input_blocks.4.1.proj_in", "P_bg12", "P_bg13"),
+            **conv("model.diffusion_model.input_blocks.4.1.proj_out", "P_bg14", "P_bg15"),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_k", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_out.0", "P_bg16", "P_bg17", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_q", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_v", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_k", "P_bg18", "P_bg19", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_out.0", "P_bg20", "P_bg21", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_q", "P_bg20", "P_bg21", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_v", "P_bg18", "P_bg19", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.ff.net.0.proj", "P_bg22", "P_bg23", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.ff.net.2", "P_bg24", "P_bg25", bias=True),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm1", "P_bg26"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm2", "P_bg26"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm3", "P_bg26"),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_k", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_out.0", "P_bg27", "P_bg28", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_q", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_v", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_k", "P_bg29", "P_bg30", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_out.0", "P_bg31", "P_bg32", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_q", "P_bg31", "P_bg32", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_v", "P_bg33", "P_bg34", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.ff.net.0.proj", "P_bg35", "P_bg36", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.ff.net.2", "P_bg37", "P_bg38", bias=True),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm1", "P_bg39"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm2", "P_bg39"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm3", "P_bg39"),
+            **easyblock("model.diffusion_model.input_blocks.5.0", "P_bg40", "P_bg41"),
+            **norm("model.diffusion_model.input_blocks.5.1.norm", "P_bg42"),
+            **conv("model.diffusion_model.input_blocks.5.1.proj_in", "P_bg43", "P_bg44"),
+            **conv("model.diffusion_model.input_blocks.5.1.proj_out", "P_bg45", "P_bg46"),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_k", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_out.0", "P_bg47", "P_bg48", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_q", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_v", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_k", "P_bg49", "P_bg50", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_out.0", "P_bg51", "P_bg52", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_q", "P_bg53", "P_bg54", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_v", "P_bg55", "P_bg56", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.ff.net.0.proj", "P_bg57", "P_bg58", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.ff.net.2", "P_bg59", "P_bg60", bias=True),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm1", "P_bg61"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm2", "P_bg61"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm3", "P_bg61"),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_k", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_out.0", "P_bg62", "P_bg63", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_q", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_v", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_k", "P_bg64", "P_bg65", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_out.0", "P_bg66", "P_bg67", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_q", "P_bg68", "P_bg69", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_v", "P_bg70", "P_bg71", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.ff.net.0.proj", "P_bg72", "P_bg73", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.ff.net.2", "P_bg74", "P_bg75", bias=True),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm1", "P_bg76"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm2", "P_bg76"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm3", "P_bg76"),
+            **conv("model.diffusion_model.input_blocks.6.0.op", "P_bg77", "P_bg78"),
+            **easyblock("model.diffusion_model.input_blocks.7.0", "P_bg78", "P_bg79"),
+            **conv("model.diffusion_model.input_blocks.7.0.skip_connection", "P_bg80", "P_bg81"),
+            **norm("model.diffusion_model.input_blocks.7.1.norm", "P_bg82"),
+            **conv("model.diffusion_model.input_blocks.7.1.proj_in", "P_bg82", "P_bg83"),
+            **conv("model.diffusion_model.input_blocks.7.1.proj_out", "P_bg84", "P_bg85"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_k", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_out.0", "P_bg86", "P_bg87", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_q", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_v", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_k", "P_bg88", "P_bg89", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_out.0", "P_bg90", "P_bg91", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_q", "P_bg92", "P_bg93", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_v", "P_bg94", "P_bg95", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.ff.net.0.proj", "P_bg96", "P_bg97", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.ff.net.2", "P_bg98", "P_bg99", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm1", "P_bg100"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm2", "P_bg100"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm3", "P_bg100"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_k", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_out.0", "P_bg101", "P_bg102", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_q", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_v", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_k", "P_bg103", "P_bg104", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_out.0", "P_bg105", "P_bg106", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_q", "P_bg107", "P_bg108", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_v", "P_bg109", "P_bg110", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.ff.net.0.proj", "P_bg111", "P_bg112", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.ff.net.2", "P_bg113", "P_bg114", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm1", "P_bg115"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm2", "P_bg115"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm3", "P_bg115"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_k", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_out.0", "P_bg116", "P_bg117", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_q", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_v", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_k", "P_bg118", "P_bg119", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_out.0", "P_bg120", "P_bg121", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_q", "P_bg122", "P_bg123", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_v", "P_bg124", "P_bg125", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.ff.net.0.proj", "P_bg126", "P_bg127", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.ff.net.2", "P_bg128", "P_bg129", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm1", "P_bg130"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm2", "P_bg130"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm3", "P_bg130"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_k", "P_bg110", "P_bg111", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_out.0", "P_bg131", "P_bg132", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_q", "P_bg131", "P_bg132", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_v", "P_bg131", "P_bg132", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_k", "P_bg133", "P_bg134", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_out.0", "P_bg135", "P_bg136", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_q", "P_bg137", "P_bg138", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_v", "P_bg139", "P_bg140", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.ff.net.0.proj", "P_bg141", "P_bg142", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.ff.net.2", "P_bg143", "P_bg144", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm1", "P_bg145"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm2", "P_bg145"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm3", "P_bg145"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_k", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_out.0", "P_bg146", "P_bg147", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_q", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_v", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_k", "P_bg148", "P_bg149", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_out.0", "P_bg150", "P_bg151", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_q", "P_bg152", "P_bg153", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_v", "P_bg154", "P_bg155", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.ff.net.0.proj", "P_bg156", "P_bg157", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.ff.net.2", "P_bg158", "P_bg159", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm1", "P_bg160"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm2", "P_bg160"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm3", "P_bg160"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_k", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_out.0", "P_bg161", "P_bg162", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_q", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_v", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_k", "P_bg163", "P_bg164", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_out.0", "P_bg165", "P_bg166", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_q", "P_bg167", "P_bg168", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_v", "P_bg169", "P_bg170", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.ff.net.0.proj", "P_bg171", "P_bg172", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.ff.net.2", "P_bg173", "P_bg174", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm1", "P_bg175"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm2", "P_bg175"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm3", "P_bg175"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_k", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_out.0", "P_bg176", "P_bg177", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_q", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_v", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_k", "P_bg178", "P_bg179", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_out.0", "P_bg180", "P_bg181", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_q", "P_bg182", "P_bg183", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_v", "P_bg184", "P_bg185", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.ff.net.0.proj", "P_bg186", "P_bg187", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.ff.net.2", "P_bg188", "P_bg189", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm1", "P_bg190"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm2", "P_bg190"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm3", "P_bg190"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_k", "P_bg191", "P_bg192", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_out.0", "P_bg193", "P_bg194", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_q", "P_bg195", "P_bg196", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_v", "P_bg197", "P_bg198", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_k", "P_bg199", "P_bg200", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_out.0", "P_bg201", "P_bg202", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_q", "P_bg203", "P_bg204", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_v", "P_bg205", "P_bg206", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.ff.net.0.proj", "P_bg207", "P_bg208", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.ff.net.2", "P_bg209", "P_bg210", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm1", "P_bg211"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm2", "P_bg211"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm3", "P_bg211"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_k", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_out.0", "P_bg212", "P_bg213", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_q", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_v", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_k", "P_bg214", "P_bg215", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_out.0", "P_bg216", "P_bg217", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_q", "P_bg218", "P_bg219", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_v", "P_bg220", "P_bg221", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.ff.net.0.proj", "P_bg222", "P_bg223", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.ff.net.2", "P_bg224", "P_bg225", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm1", "P_bg226"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm2", "P_bg226"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm3", "P_bg226"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_k", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_out.0", "P_bg227", "P_bg228", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_q", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_v", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_k", "P_bg229", "P_bg230", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_out.0", "P_bg231", "P_bg232", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_q", "P_bg233", "P_bg234", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_v", "P_bg235", "P_bg236", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.ff.net.0.proj", "P_bg237", "P_bg238", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.ff.net.2", "P_bg239", "P_bg240", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm1", "P_bg241"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm2", "P_bg241"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm3", "P_bg241"),
+            **easyblock("model.diffusion_model.input_blocks.8.0", "P_bg242", "P_bg243"),
+            **norm("model.diffusion_model.input_blocks.8.1.norm", "P_bg244"),
+            **conv("model.diffusion_model.input_blocks.8.1.proj_in", "P_bg244", "P_bg245"),
+            **conv("model.diffusion_model.input_blocks.8.1.proj_out", "P_bg246", "P_bg247"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_k", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_out.0", "P_bg248", "P_bg249", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_q", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_v", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_k", "P_bg250", "P_bg251", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_out.0", "P_bg252", "P_bg253", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_q", "P_bg254", "P_bg255", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_v", "P_bg256", "P_bg257", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.ff.net.0.proj", "P_bg258", "P_bg259", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.ff.net.2", "P_bg260", "P_bg261", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm1", "P_bg262"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm2", "P_bg262"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm3", "P_bg262"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_k", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_out.0", "P_bg263", "P_bg264", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_q", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_v", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_k", "P_bg265", "P_bg266", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_out.0", "P_bg267", "P_bg268", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_q", "P_bg269", "P_bg270", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_v", "P_bg271", "P_bg272", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.ff.net.0.proj", "P_bg273", "P_bg274", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.ff.net.2", "P_bg275", "P_bg276", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm1", "P_bg277"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm2", "P_bg277"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm3", "P_bg277"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_k", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_out.0", "P_bg278", "P_bg279", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_q", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_v", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_k", "P_bg280", "P_bg281", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_out.0", "P_bg282", "P_bg283", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_q", "P_bg284", "P_bg285", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_v", "P_bg286", "P_bg287", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.ff.net.0.proj", "P_bg288", "P_bg289", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.ff.net.2", "P_bg290", "P_bg291", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm1", "P_bg292"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm2", "P_bg292"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm3", "P_bg292"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_k", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_out.0", "P_bg293", "P_bg294", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_q", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_v", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_k", "P_bg295", "P_bg296", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_out.0", "P_bg297", "P_bg298", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_q", "P_bg299", "P_bg300", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_v", "P_bg301", "P_bg302", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.ff.net.0.proj", "P_bg303", "P_bg304", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.ff.net.2", "P_bg305", "P_bg306", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm1", "P_bg307"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm2", "P_bg307"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm3", "P_bg307"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_k", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_out.0", "P_bg308", "P_bg309", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_q", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_v", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_k", "P_bg310", "P_bg311", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_out.0", "P_bg312", "P_bg313", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_q", "P_bg314", "P_bg315", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_v", "P_bg316", "P_bg317", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.ff.net.0.proj", "P_bg318", "P_bg319", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.ff.net.2", "P_bg320", "P_bg321", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm1", "P_bg322"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm2", "P_bg322"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm3", "P_bg322"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_k", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_out.0", "P_bg322", "P_bg323", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_q", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_v", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_k", "P_bg324", "P_bg325", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_out.0", "P_bg326", "P_bg327", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_q", "P_bg328", "P_bg329", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_v", "P_bg330", "P_bg331", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.ff.net.0.proj", "P_bg332", "P_bg333", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.ff.net.2", "P_bg334", "P_bg335", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm1", "P_bg336"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm2", "P_bg336"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm3", "P_bg336"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_k", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_out.0", "P_bg337", "P_bg338", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_q", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_v", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_k", "P_bg339", "P_bg340", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_out.0", "P_bg341", "P_bg342", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_q", "P_bg343", "P_bg344", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_v", "P_bg345", "P_bg346", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.ff.net.0.proj", "P_bg347", "P_bg348", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.ff.net.2", "P_bg349", "P_bg350", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm1", "P_bg351"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm2", "P_bg351"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm3", "P_bg351"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_k", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_out.0", "P_bg352", "P_bg353", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_q", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_v", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_k", "P_bg354", "P_bg355", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_out.0", "P_bg356", "P_bg357", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_q", "P_bg358", "P_bg359", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_v", "P_bg360", "P_bg361", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.ff.net.0.proj", "P_bg362", "P_bg363", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.ff.net.2", "P_bg364", "P_bg365", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm1", "P_bg366"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm2", "P_bg366"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm3", "P_bg366"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_k", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_out.0", "P_bg367", "P_bg368", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_q", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_v", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_k", "P_bg369", "P_bg370", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_out.0", "P_bg371", "P_bg372", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_q", "P_bg373", "P_bg374", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_v", "P_bg375", "P_bg376", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.ff.net.0.proj", "P_bg377", "P_bg378", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.ff.net.2", "P_bg379", "P_bg380", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm1", "P_bg381"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm2", "P_bg381"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm3", "P_bg381"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_k", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_out.0", "P_bg382", "P_bg383", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_q", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_v", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_k", "P_bg384", "P_bg385", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_out.0", "P_bg386", "P_bg387", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_q", "P_bg388", "P_bg389", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_v", "P_bg390", "P_bg391", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.ff.net.0.proj", "P_bg392", "P_bg393", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.ff.net.2", "P_bg394", "P_bg395", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm1", "P_bg396"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm2", "P_bg396"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm3", "P_bg396"),
+            **dense("model.diffusion_model.label_emb.0.0", "P_bg397", "P_bg398", bias=True),
+            **dense("model.diffusion_model.label_emb.0.2", "P_bg399", "P_bg400", bias=True),
+            **easyblock("model.diffusion_model.middle_block.0", "P_bg401", "P_bg402"),
+            **norm("model.diffusion_model.middle_block.1.norm", "P_bg403"),
+            **conv("model.diffusion_model.middle_block.1.proj_in", "P_bg403", "P_bg404"),
+            **conv("model.diffusion_model.middle_block.1.proj_out", "P_bg405", "P_bg406"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_k", "P_bg407", "P_bg408", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_out.0", "P_bg407", "P_bg408", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_q", "P_bg407", "P_bg408", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_v", "P_bg407", "P_bg408", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_k", "P_bg409", "P_bg410", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_out.0", "P_bg411", "P_bg412", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_q", "P_bg413", "P_bg414", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_v", "P_bg415", "P_bg416", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.ff.net.0.proj", "P_bg417", "P_bg418", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.ff.net.2", "P_bg419", "P_bg420", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm1", "P_bg421"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm2", "P_bg421"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm3", "P_bg421"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_k", "P_bg422", "P_bg423", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_out.0", "P_bg422", "P_bg423", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_q", "P_bg422", "P_bg423", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_v", "P_bg422", "P_bg423", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_k", "P_bg424", "P_bg425", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_out.0", "P_bg426", "P_bg427", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_q", "P_bg428", "P_bg429", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_v", "P_bg430", "P_bg431", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.ff.net.0.proj", "P_bg432", "P_bg433", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.ff.net.2", "P_bg434", "P_bg435", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm1", "P_bg436"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm2", "P_bg436"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm3", "P_bg436"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_k", "P_bg437", "P_bg438", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_out.0", "P_bg437", "P_bg438", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_q", "P_bg437", "P_bg438", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_v", "P_bg437", "P_bg438", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_k", "P_bg439", "P_bg440", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_out.0", "P_bg441", "P_bg442", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_q", "P_bg443", "P_bg444", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_v", "P_bg445", "P_bg446", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.ff.net.0.proj", "P_bg447", "P_bg448", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.ff.net.2", "P_bg449", "P_bg450", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm1", "P_bg451"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm2", "P_bg451"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm3", "P_bg451"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_k", "P_bg452", "P_bg453", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_out.0", "P_bg452", "P_bg453", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_q", "P_bg452", "P_bg453", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_v", "P_bg452", "P_bg453", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_k", "P_bg454", "P_bg455", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_out.0", "P_bg456", "P_bg457", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_q", "P_bg458", "P_bg459", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_v", "P_bg460", "P_bg461", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.ff.net.0.proj", "P_bg462", "P_bg463", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.ff.net.2", "P_bg464", "P_bg465", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm1", "P_bg466"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm2", "P_bg466"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm3", "P_bg466"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_k", "P_bg467", "P_bg468", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_out.0", "P_bg467", "P_bg468", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_q", "P_bg467", "P_bg468", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_v", "P_bg467", "P_bg468", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_k", "P_bg469", "P_bg470", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_out.0", "P_bg471", "P_bg472", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_q", "P_bg473", "P_bg474", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_v", "P_bg475", "P_bg476", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.ff.net.0.proj", "P_bg477", "P_bg478", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.ff.net.2", "P_bg479", "P_bg480", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm1", "P_bg481"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm2", "P_bg481"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm3", "P_bg481"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_k", "P_bg482", "P_bg483", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_out.0", "P_bg482", "P_bg483", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_q", "P_bg482", "P_bg483", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_v", "P_bg482", "P_bg483", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_k", "P_bg484", "P_bg485", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_out.0", "P_bg486", "P_bg487", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_q", "P_bg488", "P_bg489", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_v", "P_bg490", "P_bg491", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.ff.net.0.proj", "P_bg492", "P_bg493", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.ff.net.2", "P_bg494", "P_bg495", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm1", "P_bg496"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm2", "P_bg496"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm3", "P_bg496"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_k", "P_bg497", "P_bg498", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_out.0", "P_bg497", "P_bg498", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_q", "P_bg497", "P_bg498", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_v", "P_bg497", "P_bg498", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_k", "P_bg499", "P_bg500", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_out.0", "P_bg501", "P_bg502", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_q", "P_bg503", "P_bg504", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_v", "P_bg505", "P_bg506", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.ff.net.0.proj", "P_bg507", "P_bg508", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.ff.net.2", "P_bg509", "P_bg510", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm1", "P_bg511"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm2", "P_bg511"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm3", "P_bg511"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_k", "P_bg512", "P_bg513", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_out.0", "P_bg512", "P_bg513", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_q", "P_bg512", "P_bg513", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_v", "P_bg512", "P_bg513", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_k", "P_bg514", "P_bg515", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_out.0", "P_bg516", "P_bg517", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_q", "P_bg518", "P_bg519", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_v", "P_bg520", "P_bg521", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.ff.net.0.proj", "P_bg522", "P_bg523", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.ff.net.2", "P_bg524", "P_bg525", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm1", "P_bg526"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm2", "P_bg526"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm3", "P_bg526"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_k", "P_bg527", "P_bg528", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_out.0", "P_bg527", "P_bg528", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_q", "P_bg527", "P_bg528", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_v", "P_bg527", "P_bg528", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_k", "P_bg529", "P_bg530", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_out.0", "P_bg531", "P_bg532", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_q", "P_bg533", "P_bg534", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_v", "P_bg535", "P_bg536", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.ff.net.0.proj", "P_bg537", "P_bg538", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.ff.net.2", "P_bg539", "P_bg540", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm1", "P_bg541"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm2", "P_bg541"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm3", "P_bg541"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_k", "P_bg542", "P_bg543", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_out.0", "P_bg542", "P_bg543", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_q", "P_bg542", "P_bg543", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_v", "P_bg542", "P_bg543", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_k", "P_bg544", "P_bg545", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_out.0", "P_bg546", "P_bg547", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_q", "P_bg548", "P_bg549", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_v", "P_bg550", "P_bg551", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.ff.net.0.proj", "P_bg552", "P_bg553", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.ff.net.2", "P_bg554", "P_bg555", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm1", "P_bg556"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm2", "P_bg556"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm3", "P_bg556"),
+            **easyblock("model.diffusion_model.middle_block.2", "P_bg557", "P_bg558"),
+            **norm("model.diffusion_model.out.0", "P_bg559"),
+            **conv("model.diffusion_model.out.2", "P_bg560", "P_bg561"),
+            **easyblock("model.diffusion_model.output_blocks.0.0", "P_bg562", "P_bg563"),
+            **conv("model.diffusion_model.output_blocks.0.0.skip_connection", "P_bg564", "P_bg565"),
+            **norm("model.diffusion_model.output_blocks.0.1.norm", "P_bg566"),
+            **conv("model.diffusion_model.output_blocks.0.1.proj_in", "P_bg566", "P_bg567"),
+            **conv("model.diffusion_model.output_blocks.0.1.proj_out", "P_bg568", "P_bg569"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_k", "P_bg570", "P_bg571", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_out.0", "P_bg570", "P_bg571", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_q", "P_bg570", "P_bg571", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_v", "P_bg570", "P_bg571", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_k", "P_bg572", "P_bg573", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_out.0", "P_bg574", "P_bg575", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_q", "P_bg576", "P_bg577", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_v", "P_bg578", "P_bg579", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.ff.net.0.proj", "P_bg580", "P_bg581", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.ff.net.2", "P_bg582", "P_bg583", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm1", "P_bg584"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm2", "P_bg584"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm3", "P_bg584"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_k", "P_bg585", "P_bg586", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_out.0", "P_bg585", "P_bg586", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_q", "P_bg585", "P_bg586", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_v", "P_bg585", "P_bg586", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_k", "P_bg587", "P_bg588", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_out.0", "P_bg589", "P_bg590", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_q", "P_bg591", "P_bg592", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_v", "P_bg593", "P_bg594", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.ff.net.0.proj", "P_bg595", "P_bg596", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.ff.net.2", "P_bg597", "P_bg598", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm1", "P_bg599"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm2", "P_bg599"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm3", "P_bg599"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_k", "P_bg600", "P_bg601", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_out.0", "P_bg600", "P_bg601", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_q", "P_bg600", "P_bg601", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_v", "P_bg600", "P_bg601", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_k", "P_bg602", "P_bg603", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_out.0", "P_bg604", "P_bg605", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_q", "P_bg606", "P_bg607", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_v", "P_bg608", "P_bg609", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.ff.net.0.proj", "P_bg610", "P_bg611", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.ff.net.2", "P_bg612", "P_bg613", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm1", "P_bg614"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm2", "P_bg614"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm3", "P_bg614"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_k", "P_bg615", "P_bg616", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_out.0", "P_bg615", "P_bg616", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_q", "P_bg615", "P_bg616", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_v", "P_bg615", "P_bg616", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_k", "P_bg617", "P_bg618", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_out.0", "P_bg619", "P_bg620", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_q", "P_bg621", "P_bg622", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_v", "P_bg623", "P_bg624", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.ff.net.0.proj", "P_bg625", "P_bg626", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.ff.net.2", "P_bg627", "P_bg628", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm1", "P_bg629"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm2", "P_bg629"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm3", "P_bg629"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_k", "P_bg630", "P_bg631", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_out.0", "P_bg630", "P_bg631", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_q", "P_bg630", "P_bg631", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_v", "P_bg630", "P_bg631", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_k", "P_bg632", "P_bg633", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_out.0", "P_bg634", "P_bg635", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_q", "P_bg636", "P_bg637", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_v", "P_bg638", "P_bg639", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.ff.net.0.proj", "P_bg640", "P_bg641", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.ff.net.2", "P_bg642", "P_bg643", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm1", "P_bg644"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm2", "P_bg644"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm3", "P_bg644"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_k", "P_bg645", "P_bg646", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_out.0", "P_bg645", "P_bg646", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_q", "P_bg645", "P_bg646", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_v", "P_bg645", "P_bg646", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_k", "P_bg647", "P_bg648", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_out.0", "P_bg649", "P_bg650", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_q", "P_bg651", "P_bg652", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_v", "P_bg653", "P_bg654", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.ff.net.0.proj", "P_bg655", "P_bg656", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.ff.net.2", "P_bg657", "P_bg658", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm1", "P_bg659"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm2", "P_bg659"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm3", "P_bg659"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_k", "P_bg660", "P_bg661", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_out.0", "P_bg660", "P_bg661", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_q", "P_bg660", "P_bg661", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_v", "P_bg660", "P_bg661", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_k", "P_bg662", "P_bg663", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_out.0", "P_bg664", "P_bg665", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_q", "P_bg666", "P_bg667", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_v", "P_bg668", "P_bg669", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.ff.net.0.proj", "P_bg670", "P_bg671", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.ff.net.2", "P_bg672", "P_bg673", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm1", "P_bg674"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm2", "P_bg674"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm3", "P_bg674"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_k", "P_bg675", "P_bg676", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_out.0", "P_bg675", "P_bg676", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_q", "P_bg675", "P_bg676", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_v", "P_bg675", "P_bg676", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_k", "P_bg677", "P_bg678", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_out.0", "P_bg679", "P_bg680", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_q", "P_bg681", "P_bg682", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_v", "P_bg683", "P_bg684", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.ff.net.0.proj", "P_bg685", "P_bg686", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.ff.net.2", "P_bg687", "P_bg688", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm1", "P_bg689"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm2", "P_bg689"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm3", "P_bg689"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_k", "P_bg690", "P_bg691", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_out.0", "P_bg690", "P_bg691", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_q", "P_bg690", "P_bg691", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_v", "P_bg690", "P_bg691", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_k", "P_bg692", "P_bg693", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_out.0", "P_bg694", "P_bg695", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_q", "P_bg696", "P_bg697", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_v", "P_bg698", "P_bg699", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.ff.net.0.proj", "P_bg700", "P_bg701", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.ff.net.2", "P_bg702", "P_bg703", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm1", "P_bg704"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm2", "P_bg704"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm3", "P_bg704"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_k", "P_bg705", "P_bg706", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_out.0", "P_bg705", "P_bg706", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_q", "P_bg705", "P_bg706", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_v", "P_bg705", "P_bg706", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_k", "P_bg707", "P_bg708", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_out.0", "P_bg709", "P_bg710", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_q", "P_bg711", "P_bg712", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_v", "P_bg713", "P_bg714", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.ff.net.0.proj", "P_bg715", "P_bg716", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.ff.net.2", "P_bg717", "P_bg718", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm1", "P_bg719"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm2", "P_bg719"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm3", "P_bg719"),
+            **easyblock("model.diffusion_model.output_blocks.1.0", "P_bg720", "P_bg721"),
+            **conv("model.diffusion_model.output_blocks.1.0.skip_connection", "P_bg722", "P_bg723"),
+            **norm("model.diffusion_model.output_blocks.1.1.norm", "P_bg724"),
+            **conv("model.diffusion_model.output_blocks.1.1.proj_in", "P_bg724", "P_bg725"),
+            **conv("model.diffusion_model.output_blocks.1.1.proj_out", "P_bg726", "P_bg727"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_k", "P_bg728", "P_bg729", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_out.0", "P_bg728", "P_bg729", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_q", "P_bg728", "P_bg729", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_v", "P_bg728", "P_bg729", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_k", "P_bg730", "P_bg731", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_out.0", "P_bg732", "P_bg733", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_q", "P_bg734", "P_bg735", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_v", "P_bg736", "P_bg737", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.ff.net.0.proj", "P_bg738", "P_bg739", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.ff.net.2", "P_bg740", "P_bg741", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm1", "P_bg742"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm2", "P_bg742"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm3", "P_bg742"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_k", "P_bg743", "P_bg744", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_out.0", "P_bg743", "P_bg744", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_q", "P_bg743", "P_bg744", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_v", "P_bg743", "P_bg744", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_k", "P_bg745", "P_bg746", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_out.0", "P_bg747", "P_bg748", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_q", "P_bg749", "P_bg750", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_v", "P_bg751", "P_bg752", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.ff.net.0.proj", "P_bg753", "P_bg754", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.ff.net.2", "P_bg755", "P_bg756", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm1", "P_bg757"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm2", "P_bg757"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm3", "P_bg757"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_k", "P_bg758", "P_bg759", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_out.0", "P_bg758", "P_bg759", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_q", "P_bg758", "P_bg759", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_v", "P_bg758", "P_bg759", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_k", "P_bg760", "P_bg761", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_out.0", "P_bg762", "P_bg763", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_q", "P_bg764", "P_bg765", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_v", "P_bg766", "P_bg767", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.ff.net.0.proj", "P_bg768", "P_bg769", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.ff.net.2", "P_bg770", "P_bg771", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm1", "P_bg772"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm2", "P_bg772"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm3", "P_bg772"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_k", "P_bg773", "P_bg774", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_out.0", "P_bg773", "P_bg774", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_q", "P_bg773", "P_bg774", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_v", "P_bg773", "P_bg774", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_k", "P_bg775", "P_bg776", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_out.0", "P_bg777", "P_bg778", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_q", "P_bg779", "P_bg780", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_v", "P_bg781", "P_bg782", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.ff.net.0.proj", "P_bg783", "P_bg784", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.ff.net.2", "P_bg785", "P_bg786", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm1", "P_bg787"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm2", "P_bg787"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm3", "P_bg787"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_k", "P_bg788", "P_bg789", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_out.0", "P_bg788", "P_bg789", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_q", "P_bg788", "P_bg789", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_v", "P_bg788", "P_bg789", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_k", "P_bg790", "P_bg791", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_out.0", "P_bg792", "P_bg793", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_q", "P_bg794", "P_bg795", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_v", "P_bg796", "P_bg797", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.ff.net.0.proj", "P_bg798", "P_bg799", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.ff.net.2", "P_bg800", "P_bg801", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm1", "P_bg802"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm2", "P_bg802"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm3", "P_bg802"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_k", "P_bg803", "P_bg804", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_out.0", "P_bg803", "P_bg804", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_q", "P_bg803", "P_bg804", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_v", "P_bg803", "P_bg804", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_k", "P_bg805", "P_bg806", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_out.0", "P_bg807", "P_bg808", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_q", "P_bg809", "P_bg810", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_v", "P_bg811", "P_bg812", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.ff.net.0.proj", "P_bg813", "P_bg814", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.ff.net.2", "P_bg815", "P_bg816", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm1", "P_bg817"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm2", "P_bg817"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm3", "P_bg817"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_k", "P_bg818", "P_bg819", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_out.0", "P_bg818", "P_bg819", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_q", "P_bg818", "P_bg819", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_v", "P_bg818", "P_bg819", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_k", "P_bg820", "P_bg821", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_out.0", "P_bg822", "P_bg823", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_q", "P_bg824", "P_bg825", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_v", "P_bg826", "P_bg827", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.ff.net.0.proj", "P_bg828", "P_bg829", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.ff.net.2", "P_bg830", "P_bg831", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm1", "P_bg832"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm2", "P_bg832"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm3", "P_bg832"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_k", "P_bg833", "P_bg834", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_out.0", "P_bg833", "P_bg834", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_q", "P_bg833", "P_bg834", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_v", "P_bg833", "P_bg834", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_k", "P_bg835", "P_bg836", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_out.0", "P_bg837", "P_bg838", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_q", "P_bg839", "P_bg840", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_v", "P_bg841", "P_bg842", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.ff.net.0.proj", "P_bg843", "P_bg844", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.ff.net.2", "P_bg845", "P_bg846", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm1", "P_bg847"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm2", "P_bg847"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm3", "P_bg847"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_k", "P_bg848", "P_bg849", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_out.0", "P_bg848", "P_bg849", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_q", "P_bg848", "P_bg849", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_v", "P_bg848", "P_bg849", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_k", "P_bg850", "P_bg851", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_out.0", "P_bg852", "P_bg853", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_q", "P_bg854", "P_bg855", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_v", "P_bg856", "P_bg857", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.ff.net.0.proj", "P_bg858", "P_bg859", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.ff.net.2", "P_bg860", "P_bg861", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm1", "P_bg862"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm2", "P_bg862"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm3", "P_bg862"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_k", "P_bg863", "P_bg864", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_out.0", "P_bg863", "P_bg864", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_q", "P_bg863", "P_bg864", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_v", "P_bg863", "P_bg864", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_k", "P_bg865", "P_bg866", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_out.0", "P_bg867", "P_bg868", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_q", "P_bg869", "P_bg870", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_v", "P_bg871", "P_bg872", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.ff.net.0.proj", "P_bg873", "P_bg874", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.ff.net.2", "P_bg875", "P_bg876", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm1", "P_bg877"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm2", "P_bg877"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm3", "P_bg877"),
+            **easyblock("model.diffusion_model.output_blocks.2.0", "P_bg878", "P_bg879"),
+            **conv("model.diffusion_model.output_blocks.2.0.skip_connection", "P_bg880", "P_bg881"),
+            **norm("model.diffusion_model.output_blocks.2.1.norm", "P_bg882"),
+            **conv("model.diffusion_model.output_blocks.2.1.proj_in", "P_bg882", "P_bg883"),
+            **conv("model.diffusion_model.output_blocks.2.1.proj_out", "P_bg884", "P_bg885"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_k", "P_bg886", "P_bg887", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_out.0", "P_bg886", "P_bg887", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_q", "P_bg886", "P_bg887", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_v", "P_bg886", "P_bg887", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_k", "P_bg888", "P_bg889", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_out.0", "P_bg890", "P_bg891", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_q", "P_bg892", "P_bg893", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_v", "P_bg894", "P_bg895", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.ff.net.0.proj", "P_bg896", "P_bg897", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.ff.net.2", "P_bg898", "P_bg899", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm1", "P_bg900"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm2", "P_bg900"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm3", "P_bg900"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_k", "P_bg901", "P_bg902", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_out.0", "P_bg901", "P_bg902", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_q", "P_bg901", "P_bg902", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_v", "P_bg901", "P_bg902", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_k", "P_bg903", "P_bg904", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_out.0", "P_bg905", "P_bg906", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_q", "P_bg907", "P_bg908", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_v", "P_bg909", "P_bg910", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.ff.net.0.proj", "P_bg911", "P_bg912", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.ff.net.2", "P_bg913", "P_bg914", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm1", "P_bg915"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm2", "P_bg915"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm3", "P_bg915"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_k", "P_bg916", "P_bg917", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_out.0", "P_bg916", "P_bg917", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_q", "P_bg916", "P_bg917", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_v", "P_bg916", "P_bg917", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_k", "P_bg918", "P_bg919", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_out.0", "P_bg920", "P_bg921", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_q", "P_bg922", "P_bg923", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_v", "P_bg924", "P_bg925", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.ff.net.0.proj", "P_bg926", "P_bg927", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.ff.net.2", "P_bg928", "P_bg929", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm1", "P_bg930"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm2", "P_bg930"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm3", "P_bg930"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_k", "P_bg931", "P_bg932", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_out.0", "P_bg931", "P_bg932", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_q", "P_bg931", "P_bg932", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_v", "P_bg931", "P_bg932", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_k", "P_bg933", "P_bg934", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_out.0", "P_bg935", "P_bg936", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_q", "P_bg937", "P_bg938", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_v", "P_bg939", "P_bg940", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.ff.net.0.proj", "P_bg941", "P_bg942", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.ff.net.2", "P_bg943", "P_bg944", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm1", "P_bg945"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm2", "P_bg945"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm3", "P_bg945"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_k", "P_bg946", "P_bg947", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_out.0", "P_bg946", "P_bg947", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_q", "P_bg946", "P_bg947", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_v", "P_bg946", "P_bg947", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_k", "P_bg948", "P_bg949", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_out.0", "P_bg950", "P_bg951", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_q", "P_bg952", "P_bg953", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_v", "P_bg954", "P_bg955", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.ff.net.0.proj", "P_bg956", "P_bg957", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.ff.net.2", "P_bg958", "P_bg959", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm1", "P_bg960"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm2", "P_bg960"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm3", "P_bg960"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_k", "P_bg961", "P_bg962", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_out.0", "P_bg961", "P_bg962", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_q", "P_bg961", "P_bg962", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_v", "P_bg961", "P_bg962", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_k", "P_bg963", "P_bg964", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_out.0", "P_bg965", "P_bg966", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_q", "P_bg967", "P_bg968", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_v", "P_bg969", "P_bg970", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.ff.net.0.proj", "P_bg971", "P_bg972", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.ff.net.2", "P_bg973", "P_bg974", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm1", "P_bg975"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm2", "P_bg975"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm3", "P_bg975"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_k", "P_bg976", "P_bg977", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_out.0", "P_bg976", "P_bg977", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_q", "P_bg976", "P_bg977", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_v", "P_bg976", "P_bg977", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_k", "P_bg978", "P_bg979", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_out.0", "P_bg980", "P_bg981", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_q", "P_bg982", "P_bg983", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_v", "P_bg984", "P_bg985", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.ff.net.0.proj", "P_bg986", "P_bg987", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.ff.net.2", "P_bg988", "P_bg989", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm1", "P_bg990"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm2", "P_bg990"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm3", "P_bg990"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_k", "P_bg991", "P_bg992", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_out.0", "P_bg991", "P_bg992", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_q", "P_bg991", "P_bg992", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_v", "P_bg991", "P_bg992", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_k", "P_bg993", "P_bg994", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_out.0", "P_bg995", "P_bg996", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_q", "P_bg997", "P_bg998", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_v", "P_bg999", "P_bg1000", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.ff.net.0.proj", "P_bg1001", "P_bg1002", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.ff.net.2", "P_bg1003", "P_bg1004", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm1", "P_bg1005"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm2", "P_bg1005"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm3", "P_bg1005"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_k", "P_bg1006", "P_bg1007", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_out.0", "P_bg1006", "P_bg1007", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_q", "P_bg1006", "P_bg1007", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_v", "P_bg1006", "P_bg1007", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_k", "P_bg1008", "P_bg1009", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_out.0", "P_bg1010", "P_bg1011", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_q", "P_bg1012", "P_bg1013", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_v", "P_bg1014", "P_bg1015", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.ff.net.0.proj", "P_bg1016", "P_bg1017", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.ff.net.2", "P_bg1018", "P_bg1019", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm1", "P_bg1020"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm2", "P_bg1020"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm3", "P_bg1020"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_k", "P_bg1021", "P_bg1022", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_out.0", "P_bg1021", "P_bg1022", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_q", "P_bg1021", "P_bg1022", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_v", "P_bg1021", "P_bg1022", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_k", "P_bg1023", "P_bg1024", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_out.0", "P_bg1025", "P_bg1026", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_q", "P_bg1027", "P_bg1028", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_v", "P_bg1029", "P_bg1030", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.ff.net.0.proj", "P_bg1031", "P_bg1032", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.ff.net.2", "P_bg1033", "P_bg1034", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm1", "P_bg1035"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm2", "P_bg1035"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm3", "P_bg1035"),
+            **conv("model.diffusion_model.output_blocks.2.2.conv", "P_bg1036", "P_bg1037"),
+            **easyblock("model.diffusion_model.output_blocks.3.0", "P_bg1038", "P_bg1039"),
+            **conv("model.diffusion_model.output_blocks.3.0.skip_connection", "P_bg1040", "P_bg1041"),
+            **norm("model.diffusion_model.output_blocks.3.1.norm", "P_bg1042"),
+            **conv("model.diffusion_model.output_blocks.3.1.proj_in", "P_bg1042", "P_bg1043"),
+            **conv("model.diffusion_model.output_blocks.3.1.proj_out", "P_bg1044", "P_bg1045"),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_k", "P_bg1046", "P_bg1047", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_out.0", "P_bg1046", "P_bg1047", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_q", "P_bg1046", "P_bg1047", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_v", "P_bg1046", "P_bg1047", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_k", "P_bg1048", "P_bg1049", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_out.0", "P_bg1050", "P_bg1051", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_q", "P_bg1052", "P_bg1053", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_v", "P_bg1054", "P_bg1055", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.ff.net.0.proj", "P_bg1056", "P_bg1057", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.ff.net.2", "P_bg1058", "P_bg1059", bias=True),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm1", "P_bg1060"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm2", "P_bg1060"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm3", "P_bg1060"),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_k", "P_bg1061", "P_bg1062", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_out.0", "P_bg1061", "P_bg1062", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_q", "P_bg1061", "P_bg1062", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_v", "P_bg1061", "P_bg1062", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_k", "P_bg1063", "P_bg1064", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_out.0", "P_bg1065", "P_bg1066", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_q", "P_bg1067", "P_bg1068", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_v", "P_bg1069", "P_bg1070", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.ff.net.0.proj", "P_bg1071", "P_bg1072", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.ff.net.2", "P_bg1073", "P_bg1074", bias=True),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm1", "P_bg1075"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm2", "P_bg1075"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm3", "P_bg1075"),
+            **easyblock("model.diffusion_model.output_blocks.4.0", "P_bg1076", "P_bg1077"),
+            **conv("model.diffusion_model.output_blocks.4.0.skip_connection", "P_bg1078", "P_bg1079"),
+            **norm("model.diffusion_model.output_blocks.4.1.norm", "P_bg1080"),
+            **conv("model.diffusion_model.output_blocks.4.1.proj_in", "P_bg1080", "P_bg1081"),
+            **conv("model.diffusion_model.output_blocks.4.1.proj_out", "P_bg1082", "P_bg1083"),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_k", "P_bg1084", "P_bg1085", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_out.0", "P_bg1084", "P_bg1085", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_q", "P_bg1084", "P_bg1085", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_v", "P_bg1084", "P_bg1085", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_k", "P_bg1086", "P_bg1087", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_out.0", "P_bg1088", "P_bg1089", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_q", "P_bg1090", "P_bg1091", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_v", "P_bg1092", "P_bg1093", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.ff.net.0.proj", "P_bg1094", "P_bg1095", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.ff.net.2", "P_bg1096", "P_bg1097", bias=True),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm1", "P_bg1098"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm2", "P_bg1098"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm3", "P_bg1098"),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_k", "P_bg1099", "P_bg1100", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_out.0", "P_bg1099", "P_bg1100", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_q", "P_bg1099", "P_bg1100", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_v", "P_bg1099", "P_bg1100", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_k", "P_bg1101", "P_bg1102", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_out.0", "P_bg1103", "P_bg1104", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_q", "P_bg1105", "P_bg1106", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_v", "P_bg1107", "P_bg1108", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.ff.net.0.proj", "P_bg1109", "P_bg1110", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.ff.net.2", "P_bg1111", "P_bg1112", bias=True),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm1", "P_bg1113"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm2", "P_bg1113"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm3", "P_bg1113"),
+            **easyblock("model.diffusion_model.output_blocks.5.0", "P_bg1114", "P_bg1115"),
+            **conv("model.diffusion_model.output_blocks.5.0.skip_connection", "P_bg1116", "P_bg1117"),
+            **norm("model.diffusion_model.output_blocks.5.1.norm", "P_bg1118"),
+            **conv("model.diffusion_model.output_blocks.5.1.proj_in", "P_bg1118", "P_bg1119"),
+            **conv("model.diffusion_model.output_blocks.5.1.proj_out", "P_bg1120", "P_bg1121"),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_k", "P_bg1122", "P_bg1123", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_out.0", "P_bg1122", "P_bg1123", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_q", "P_bg1122", "P_bg1123", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_v", "P_bg1122", "P_bg1123", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_k", "P_bg1124", "P_bg1125", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_out.0", "P_bg1126", "P_bg1127", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_q", "P_bg1128", "P_bg1129", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_v", "P_bg1130", "P_bg1131", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.ff.net.0.proj", "P_bg1132", "P_bg1133", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.ff.net.2", "P_bg1134", "P_bg1135", bias=True),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm1", "P_bg1136"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm2", "P_bg1136"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm3", "P_bg1136"),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_k", "P_bg1137", "P_bg1138", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_out.0", "P_bg1137", "P_bg1138", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_q", "P_bg1137", "P_bg1138", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_v", "P_bg1137", "P_bg1138", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_k", "P_bg1139", "P_bg1140", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_out.0", "P_bg1141", "P_bg1142", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_q", "P_bg1143", "P_bg1144", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_v", "P_bg1145", "P_bg1146", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.ff.net.0.proj", "P_bg1147", "P_bg1148", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.ff.net.2", "P_bg1149", "P_bg1150", bias=True),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm1", "P_bg1151"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm2", "P_bg1151"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm3", "P_bg1151"),
+            **conv("model.diffusion_model.output_blocks.5.2.conv", "P_bg1152", "P_bg1153"),
+            **easyblock("model.diffusion_model.output_blocks.6.0", "P_bg1154", "P_bg1155"),
+            **conv("model.diffusion_model.output_blocks.6.0.skip_connection", "P_bg1156", "P_bg1157"),
+            **easyblock("model.diffusion_model.output_blocks.7.0", "P_bg1158", "P_bg1159"),
+            **conv("model.diffusion_model.output_blocks.7.0.skip_connection", "P_bg1160", "P_bg1161"),
+            **easyblock("model.diffusion_model.output_blocks.8.0", "P_bg1162", "P_bg1163"),
+            **conv("model.diffusion_model.output_blocks.8.0.skip_connection", "P_bg1164", "P_bg1165"),
+            **dense("model.diffusion_model.time_embed.0", "P_bg1166", "P_bg1167", bias=True),
+            **dense("model.diffusion_model.time_embed.2", "P_bg1168", "P_bg1169", bias=True),
+            # Text Encoder 1
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.position_embedding", "P_bg1170", "P_bg1171", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.token_embedding", "P_bg1172", "P_bg1173", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc1", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc2", "P_bg1178", "P_bg1179", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.k_proj", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.out_proj", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.q_proj", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.v_proj", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc1", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc2", "P_bg1186", "P_bg1187", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.k_proj", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.out_proj", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.q_proj", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.v_proj", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc1", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc2", "P_bg1194", "P_bg1195", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.k_proj", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.out_proj", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.q_proj", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.v_proj", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc1", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc2", "P_bg1202", "P_bg1203", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.k_proj", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.out_proj", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.q_proj", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.v_proj", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc1", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc2", "P_bg1210", "P_bg1211", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.k_proj", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.out_proj", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.q_proj", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.v_proj", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc1", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc2", "P_bg1218", "P_bg1219", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.k_proj", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.out_proj", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.q_proj", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.v_proj", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc1", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc2", "P_bg1226", "P_bg1227", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.k_proj", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.out_proj", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.q_proj", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.v_proj", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc1", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc2", "P_bg1234", "P_bg1235", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.k_proj", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.out_proj", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.q_proj", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.v_proj", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc1", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc2", "P_bg1242", "P_bg1243", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.k_proj", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.out_proj", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.q_proj", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.v_proj", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc1", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc2", "P_bg1250", "P_bg1251", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.k_proj", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.out_proj", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.q_proj", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.v_proj", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc1", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc2", "P_bg1258", "P_bg1259", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.k_proj", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.out_proj", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.q_proj", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.v_proj", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc1", "P_bg1264", "P_bg1265", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc2", "P_bg1266", "P_bg1267", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.k_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.out_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.q_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.v_proj", "P_bg1268", "P_bg1269", bias=False),
+            # Text Encoder 2
+            **dense("conditioner.embedders.1.model.token_embedding", "P_bg1272", "P_bg1273", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1278", "P_bg1279", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1284", "P_bg1285", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1290", "P_bg1291", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1296", "P_bg1297", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1302", "P_bg1303", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1308", "P_bg1309", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1314", "P_bg1315", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1320", "P_bg1321", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1326", "P_bg1327", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1332", "P_bg1333", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1338", "P_bg1339", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1344", "P_bg1345", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1350", "P_bg1351", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1356", "P_bg1357", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1362", "P_bg1363", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1368", "P_bg1369", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1374", "P_bg1375", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1380", "P_bg1381", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1386", "P_bg1387", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1392", "P_bg1393", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1398", "P_bg1399", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1404", "P_bg1405", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1410", "P_bg1411", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1416", "P_bg1417", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1422", "P_bg1423", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1428", "P_bg1429", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1434", "P_bg1435", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1440", "P_bg1441", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1446", "P_bg1447", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1452", "P_bg1453", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1458", "P_bg1459", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1460", "P_bg1461", bias=False),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj", "P_bg1460", "P_bg1461", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1460", "P_bg1461", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1464", "P_bg1465", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1466", "P_bg1467", bias=False),
+        }
+    )
